@@ -1,23 +1,23 @@
 {
-  "name": "glitch-hello-eleventy",
+  "name": "Chimera Collective Webring",
   "version": "0.1.0",
-  "description": "A simple blog, set up with Eleventy.",
+  "description": "A simple web ring, running on Eleventy.",
+	"type": "module",
   "scripts": {
     "start": "eleventy --serve",
     "build": "NODE_ENV=production eleventy"
   },
   "dependencies": {
+    "@11ty/eleventy": "^3.0.0-alpha.20",
+    "@11ty/eleventy-img": "^4.0.2",
     "eleventy-plugin-seo": "0.5.2",
     "luxon": "^3.4.0"
   },
-  "devDependencies": {
-    "@11ty/eleventy": "^2.0.0"
-  },
   "engines": {
-    "node": "14"
+    "node": "20"
   },
   "repository": {
-    "url": "https://glitch.com/edit/#!/glitch-hello-eleventy"
+    "url": "https://github.com/heatloss/chimera"
   },
   "glitch": {
     "projectType": "generated_static"
